updated migration

+++ db/migrate/20121022220338_add_item_id_to_line_items.rb
@@ -1,5 +1,5 @@
 class AddItemIdToLineItems < ActiveRecord::Migration
   def change
-    add_column :line_items, :item_id, :integer
+    add_column :line_items, :item_id, :float
   end
 end
